Setup front-end home page & user input page
There are two buttons sell and buy on the home page, both link to the
input page. In the final version, the dining halls should be dropdown
menus, but I use datalist to do it for now to keep it simple. The two
pairs of upper/lower bounds alerts when lower > upper, but this should
be more user-friendly in the future.