Add match and waitingpage
When the user click on submit button to submit request/offer, a confirm
message is displayed and a “find me a match” button is shown. Clicking
on the button leads the user to the waiting page, after 3
second(backend should be calculating at that time), it goes to match
page, on which match info is displayed.